src/generate_keyphrases/requirements.txt
flask
dapr
cloudevents
uvicorn
typing-extensions
azure-ai-textanalytics
azure-identity
tenacity==8.2.2
